requirements.txt
Flask
Flask-SQLAlchemy
ffmpeg-python
werkzeug